parser: improve qualified names
First off, I renamed the `field_access` rule to `qualified_name` since
it really designates more generally qualified names, not just fields
(e.g. class names and table names qualified by the database they're in);
I removed the `table` and `field` fields from the rule accordingly due
to the change of semantics.

I then changed the content of the rule so that the part starting from
the first dot isn't just parsed as a single `field_identifier` node, but
reuses the `identifier` and parses each dot as an individual node. This
has two advantages:

- it puts the dots into the syntax tree as anonymous nodes, available
  for querying by syntax highlighting tools and the like;

- it outputs as many `identifier` nodes into the AST as there are name
  in the qualified name, which is closer to how the code is actually
  parsed in ABL.

Hopefully this is less hacky, though it involves the addition of an
external scanner to do the lookahead that allows us to tell whether a
dot is a statement terminator or part of a qualified name ("namedot").
The code is straightforward C and involves no dynamic memory allocation.