parser: break up keyword in some phrases
This makes the parser less rigid, allowing for comments, line-breaks,
includes and multiple white space between the keyword as well as
outputting a syntax tree that's more faithful to the actual content of
the code.

I also took the opportunity to make the "PROCEDURE" keyword of the block
terminator of the PROCEDURE statement optional and to remove the
useless space after the "AVAIL[ABLE]" keyword.